refactor: :recycle: move standards conversion logic into embed-xml helper
